add confirm

+++ src/app/item-page/item-page.component.ts
@@ -10,38 +10,42 @@ import { HttpErrorResponse } from '@angular/common/http';
   styleUrl: './item-page.component.css',
 })
 export class ItemPageComponent {
-  constructor(private location: Location, private itemService: ItemService) { }
+  constructor(private location: Location, private itemService: ItemService) {}
   @Input() result: Item = {
-    id: "",
+    id: '',
     description: '',
-    done: false
+    done: false,
   };
 
   onDeleteClick(): void {
-    console.log('delete');
-    this.itemService.deleteItem(this.result.id);
+    const confirmDelete = confirm('Do you want to delete?');
+    if (confirmDelete === true) {
+      console.log('delete');
+      this.itemService.deleteItem(this.result.id);
+    }
   }
 
   onSaveClick(): void {
     console.log('onSaveClick');
     console.log(this.result);
-    if (this.result.id) {
-          this.itemService.updateItem(this.result).subscribe(
-            (data: Item) => {
-              console.log('Item updated successfully:', data);
-            },
-            (error: HttpErrorResponse) => {
-              console.log('Item updated  response error');
-              console.log(error);
-            }
-          );
-    } else {
-      console.log("create");
-      this.itemService.createItem(this.result);
+    const confirmSave = confirm('Do you want to save?');
+    if (confirmSave === true) {
+      if (this.result.id) {
+        console.log('update');
+        this.itemService.updateItem(this.result).subscribe(
+          (data: Item) => {
+            console.log('Item updated successfully:', data);
+          },
+          (error: HttpErrorResponse) => {
+            console.log('Item updated  response error');
+            console.log(error);
+          }
+        );
+      } else {
+        console.log('create');
+        this.itemService.createItem(this.result);
+      }
     }
-
-
-
   }
   onBackClick(): void {
     this.location.back();
